requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
gpiod==2.2.3
gpiodevice==0.0.5
idna==3.10
inky>=1.5.0
numpy==2.2.1
Pillow>=10.0.0
python-dotenv>=1.0.0
pytz==2024.2
requests>=2.31.0
smbus2==0.5.0
spidev==3.6
urllib3==2.3.0
